pyproject.toml
[project]
name = "gs-soccer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"

dependencies = [
    "torch",
    "torchvision",
    "gymnasium[classic-control]",
    "genesis-world>=0.3.14",
    "tensorboard>=2.20.0",
    "skrl[torch]>=1.4.3",
    "ipython>=9.10.0",
    "setuptools<80.0.0",
    "rsl-rl-lib >= 2.2.1",
]

[project.optional-dependencies]
cuda = ["torch", "torchvision", "triton"]
rocm = ["torch", "torchvision", "triton", "triton-kernels"]

[tool.uv]
override-dependencies = [
    "torch==2.9.1+rocm7.2.0; extra == 'rocm'",
    "torchvision==0.25.0+rocm7.2.0; extra == 'rocm'",
    "triton==3.4.0+rocm7.2.0; extra == 'rocm'",
    "triton-kernels; extra == 'rocm'",
    "torch; extra == 'cuda'",
    "torchvision; extra == 'cuda'",
    "triton; extra == 'cuda'",
]

conflicts = [[{ extra = "cuda" }, { extra = "rocm" }]]

[tool.uv.sources]
torch = [
  { url = "https://repo.radeon.com/rocm/manylinux/rocm-rel-7.2/torch-2.9.1%2Brocm7.2.0.lw.git7e1940d4-cp312-cp312-linux_x86_64.whl", extra = "rocm", marker = "extra == 'rocm'" },
  { index = "pytorch-cuda130", extra = "cuda", marker = "extra == 'cuda'" }
]
torchvision = [
  { url = "https://repo.radeon.com/rocm/manylinux/rocm-rel-7.2/torchvision-0.24.0%2Brocm7.2.0.gitb919bd0c-cp312-cp312-linux_x86_64.whl", extra = "rocm", marker = "extra == 'rocm'" },
  { index = "pytorch-cuda130", extra = "cuda", marker = "extra == 'cuda'" }
]
triton = [
  { url = "https://repo.radeon.com/rocm/manylinux/rocm-rel-7.2/triton-3.4.0%2Brocm7.2.0.git0cace8d2-cp312-cp312-linux_x86_64.whl", extra = "rocm", marker = "extra == 'rocm'" },
  { index = "pytorch-cuda130", extra = "cuda", marker = "extra == 'cuda'" }
]
triton-kernels = [
  { url = "https://repo.radeon.com/rocm/manylinux/rocm-rel-7.2/triton_kernels-1.0.0-py3-none-any.whl", extra = "rocm", marker = "extra == 'rocm'" },
]

[[tool.uv.index]]
name = "pytorch-cuda130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true


